Break up giant org spec into functional parts Other odds and ends
