+++ spec/models/import_organisations_spec.rb
@@ -1,0 +1,92 @@
+require 'rails_helper'
+describe Organisation, :type => :model do
+
+  it 'can humanize with all first capitals' do
+    expect("HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW".humanized_all_first_capitals).to eq("Harrow Baptist Church, College Road, Harrow")
+  end
+
+  describe 'Creating of Organisations from CSV file' do
+    before(:all){ @headers = 'Title,Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')}
+
+    it 'must not override an existing organisation' do
+      FactoryGirl.create(:organisation, :email => nil,  :name => 'Indian Elders Association',
+                        :description => 'Care for the elderly', :address => '64 pinner road', :postcode => 'HA1 3RE', :donation_info => 'www.indian-elders.co.uk/donate')
+      fields = CSV.parse('INDIAN ELDERS ASSOCIATION,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH,COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      org = create_organisation(fields)
+      expect(org).to be_nil
+    end
+
+    it 'must not create org when date removed is not nil' do
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,2009-05-28,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      org = create_organisation(fields)
+      expect(org).to be_nil
+    end
+
+    it 'must be able to generate multiple Organisations from text file' do
+      expect{
+        Organisation.import_addresses 'db/data.csv', 2
+      }.to change(Organisation, :count).by 2
+    end
+
+    it 'must fail gracefully when encountering error in generating multiple Organisations from text file' do
+      attempted_number_to_import = 1006
+      actual_number_to_import = 642
+      allow(Organisation).to receive(:create_from_array).and_raise(CSV::MalformedCSVError)
+      expect(lambda {
+        Organisation.import_addresses 'db/data.csv', attempted_number_to_import
+      }).to change(Organisation, :count).by(0)
+    end
+
+    it 'must be able to handle no postcode in text representation' do
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      org = create_organisation(fields)
+      expect(org.name).to eq('Harrow Baptist Church')
+      expect(org.description).to eq('No information recorded')
+      expect(org.address).to eq('Harrow Baptist Church, College Road, Harrow')
+      expect(org.postcode).to eq('')
+      expect(org.website).to eq('http://www.harrow-baptist.org.uk')
+      expect(org.telephone).to eq('020 8863 7837')
+      expect(org.donation_info).to eq(nil)
+    end
+
+    it 'must be able to handle no address in text representation' do
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,,http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      org = create_organisation(fields)
+      expect(org.name).to eq('Harrow Baptist Church')
+      expect(org.description).to eq('No information recorded')
+      expect(org.address).to eq('')
+      expect(org.postcode).to eq('')
+      expect(org.website).to eq('http://www.harrow-baptist.org.uk')
+      expect(org.telephone).to eq('020 8863 7837')
+      expect(org.donation_info).to eq(nil)
+    end
+
+    it 'must be able to generate Organisation from text representation ensuring words in correct case and postcode is extracted from address' do
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      org = create_organisation(fields)
+      expect(org.name).to eq('Harrow Baptist Church')
+      expect(org.description).to eq('No information recorded')
+      expect(org.address).to eq('Harrow Baptist Church, College Road, Harrow')
+      expect(org.postcode).to eq('HA1 1BA')
+      expect(org.website).to eq('http://www.harrow-baptist.org.uk')
+      expect(org.telephone).to eq('020 8863 7837')
+      expect(org.donation_info).to eq(nil)
+    end
+
+    it 'should raise error if no columns found' do
+      #Headers are without Title header
+      @headers = 'Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')
+      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
+      expect(lambda{
+        org = create_organisation(fields)
+      }).to raise_error
+    end
+
+    def create_organisation(fields)
+      row = CSV::Row.new(@headers, fields.flatten)
+      Organisation.create_from_array(row, true)
+    end
+
+
+  end
+end
