+++ spec/models/organisation_spec.rb
@@ -11,8 +11,7 @@ describe Organisation, :type => :model do
     @category3 = FactoryGirl.create(:category, :charity_commission_id => 108)
     @category4 = FactoryGirl.create(:category, :charity_commission_id => 302)
     @category5 = FactoryGirl.create(:category, :charity_commission_id => 306)
-    @org1 = FactoryGirl.build(:organisation, :email => nil, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate')
-    @org1.save!
+    @org1 = FactoryGirl.create(:organisation, :email => nil, :name => 'Harrow Bereavement Counselling', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate')
     @org2 = FactoryGirl.build(:organisation, :email => nil,  :name => 'Indian Elders Association',
                               :description => 'Care for the elderly', :address => '64 pinner road', :postcode => 'HA1 3RE', :donation_info => 'www.indian-elders.co.uk/donate')
     @org2.categories << @category1
@@ -249,164 +248,6 @@ describe Organisation, :type => :model do
     expect(@org1).to respond_to(:users)
   end
 
-  it 'can humanize with all first capitals' do
-    expect("HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW".humanized_all_first_capitals).to eq("Harrow Baptist Church, College Road, Harrow")
-  end
-
-  describe 'Creating of Organisations from CSV file' do
-    before(:all){ @headers = 'Title,Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')}
-
-    it 'must not override an existing organisation' do
-      fields = CSV.parse('INDIAN ELDERS ASSOCIATION,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH,COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      org = create_organisation(fields)
-      expect(org).to be_nil
-    end
-
-    it 'must not create org when date removed is not nil' do
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,2009-05-28,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      org = create_organisation(fields)
-      expect(org).to be_nil
-    end
-
-    it 'must be able to generate multiple Organisations from text file' do
-      expect{
-        Organisation.import_addresses 'db/data.csv', 2
-      }.to change(Organisation, :count).by 2
-    end
-
-    it 'must fail gracefully when encountering error in generating multiple Organisations from text file' do
-      attempted_number_to_import = 1006
-      actual_number_to_import = 642
-      allow(Organisation).to receive(:create_from_array).and_raise(CSV::MalformedCSVError)
-      expect(lambda {
-        Organisation.import_addresses 'db/data.csv', attempted_number_to_import
-      }).to change(Organisation, :count).by(0)
-    end
-
-    it 'must be able to handle no postcode in text representation' do
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      org = create_organisation(fields)
-      expect(org.name).to eq('Harrow Baptist Church')
-      expect(org.description).to eq('No information recorded')
-      expect(org.address).to eq('Harrow Baptist Church, College Road, Harrow')
-      expect(org.postcode).to eq('')
-      expect(org.website).to eq('http://www.harrow-baptist.org.uk')
-      expect(org.telephone).to eq('020 8863 7837')
-      expect(org.donation_info).to eq(nil)
-    end
-
-    it 'must be able to handle no address in text representation' do
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,,http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      org = create_organisation(fields)
-      expect(org.name).to eq('Harrow Baptist Church')
-      expect(org.description).to eq('No information recorded')
-      expect(org.address).to eq('')
-      expect(org.postcode).to eq('')
-      expect(org.website).to eq('http://www.harrow-baptist.org.uk')
-      expect(org.telephone).to eq('020 8863 7837')
-      expect(org.donation_info).to eq(nil)
-    end
-
-    it 'must be able to generate Organisation from text representation ensuring words in correct case and postcode is extracted from address' do
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      org = create_organisation(fields)
-      expect(org.name).to eq('Harrow Baptist Church')
-      expect(org.description).to eq('No information recorded')
-      expect(org.address).to eq('Harrow Baptist Church, College Road, Harrow')
-      expect(org.postcode).to eq('HA1 1BA')
-      expect(org.website).to eq('http://www.harrow-baptist.org.uk')
-      expect(org.telephone).to eq('020 8863 7837')
-      expect(org.donation_info).to eq(nil)
-    end
-
-
-    it 'should raise error if no columns found' do
-      #Headers are without Title header
-      @headers = 'Charity Number,Activities,Contact Name,Contact Address,website,Contact Telephone,date registered,date removed,accounts date,spending,income,company number,OpenlyLocalURL,twitter account name,facebook account name,youtube account name,feed url,Charity Classification,signed up for 1010,last checked,created at,updated at,Removed?'.split(',')
-      fields = CSV.parse('HARROW BAPTIST CHURCH,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      expect(lambda{
-        org = create_organisation(fields)
-      }).to raise_error
-    end
-
-
-    def create_organisation(fields)
-      row = CSV::Row.new(@headers, fields.flatten)
-      Organisation.create_from_array(row, true)
-    end
-
-    context "importing category relations" do
-      let(:fields) do
-        CSV.parse('HARROW BEREAVEMENT COUNSELLING,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,"207,305,108,302,306",false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      end
-      let(:row) do
-        CSV::Row.new(@headers, fields.flatten)
-      end
-      let(:fields_cat_missing) do
-        CSV.parse('HARROW BEREAVEMENT COUNSELLING,1129832,NO INFORMATION RECORDED,MR JOHN ROSS NEWBY,"HARROW BAPTIST CHURCH, COLLEGE ROAD, HARROW, HA1 1BA",http://www.harrow-baptist.org.uk,020 8863 7837,2009-05-27,,,,,,http://OpenlyLocal.com/charities/57879-HARROW-BAPTIST-CHURCH,,,,,,false,2010-09-20T21:38:52+01:00,2010-08-22T22:19:07+01:00,2012-04-15T11:22:12+01:00,*****')
-      end
-      let(:row_cat_missing) do
-        CSV::Row.new(@headers, fields_cat_missing.flatten)
-      end
-      it 'must be able to avoid org category relations from text file when org does not exist' do
-        @org4 = FactoryGirl.build(:organisation, :name => 'Fellowship For Management In Food Distribution', :description => 'Bereavement Counselling', :address => '64 pinner road', :postcode => 'HA1 3TE', :donation_info => 'www.harrow-bereavment.co.uk/donate')
-        @org4.save!
-        [102,206,302].each do |id|
-          FactoryGirl.build(:category, :charity_commission_id => id).save!
-        end
-        attempted_number_to_import = 2
-        number_cat_org_relations_generated = 3
-        expect(lambda {
-          Organisation.import_category_mappings 'db/data.csv', attempted_number_to_import
-        }).to change(CategoryOrganisation, :count).by(number_cat_org_relations_generated)
-      end
-
-      it "allows us to import categories" do
-        org = Organisation.import_categories_from_array(row)
-        expect(org.categories.length).to eq 5
-        [207,305,108,302,306].each do |id|
-          expect(org.categories).to include(Category.find_by_charity_commission_id(id))
-        end
-      end
-
-      it 'must fail gracefully when encountering error in importing categories from text file' do
-        attempted_number_to_import = 2
-        allow(Organisation).to receive(:import_categories_from_array).and_raise(CSV::MalformedCSVError)
-        expect(lambda {
-          Organisation.import_category_mappings 'db/data.csv', attempted_number_to_import
-        }).to change(Organisation, :count).by(0)
-      end
-
-      it "should import categories when matching org is found" do
-        expect(Organisation).to receive(:check_columns_in).with(row)
-        expect(Organisation).to receive(:find_by_name).with('Harrow Bereavement Counselling').and_return @org1
-        array = double('Array')
-        [{:cc_id => 207, :cat => @cat1}, {:cc_id => 305, :cat => @cat2}, {:cc_id => 108, :cat => @cat3},
-         {:cc_id => 302, :cat => @cat4}, {:cc_id => 306, :cat => @cat5}]. each do |cat_hash|
-          expect(Category).to receive(:find_by_charity_commission_id).with(cat_hash[:cc_id]).and_return(cat_hash[:cat])
-          expect(array).to receive(:<<).with(cat_hash[:cat])
-        end
-        expect(@org1).to receive(:categories).exactly(5).times.and_return(array)
-        org = Organisation.import_categories_from_array(row)
-        expect(org).not_to be_nil
-      end
-
-      it "should not import categories when no matching organisation" do
-        expect(Organisation).to receive(:check_columns_in).with(row)
-        expect(Organisation).to receive(:find_by_name).with('Harrow Bereavement Counselling').and_return nil
-        org = Organisation.import_categories_from_array(row)
-        expect(org).to be_nil
-      end
-
-      it "should not import categories when none are listed" do
-        expect(Organisation).to receive(:check_columns_in).with(row_cat_missing)
-        expect(Organisation).to receive(:find_by_name).with('Harrow Bereavement Counselling').and_return @org1
-        org = Organisation.import_categories_from_array(row_cat_missing)
-        expect(org).not_to be_nil
-      end
-    end
-  end
-
   it 'should geocode when address changes' do
     new_address = '60 pinner road'
     expect(@org1).to receive(:geocode)
@@ -420,55 +261,7 @@ describe Organisation, :type => :model do
     expect(org).to receive(:geocode)    
     org.save
   end
-  
-  # not sure if we need SQL injection security tests like this ...
-  # org = Organisation.new(:address =>"blah", :gmaps=> ";DROP DATABASE;")
-  # org = Organisation.new(:address =>"blah", :name=> ";DROP DATABASE;")
 
-  describe "importing emails" do
-    it "should have a method import_emails" do
-      filename = "db/emails.csv"
-      expect(Organisation).to receive(:add_email).twice.and_call_original
-      expect(Organisation).to receive(:import).with(filename,2,false).and_call_original
-     # do |&arg|
-     #   byebug
-     #   Organisation.add_email(&arg)
-     # end
-      Organisation.import_emails(filename,2,false)
-    end
-
-    it 'should handle absence of org gracefully' do
-      expect(Organisation).to receive(:where).with("UPPER(name) LIKE ? ", "%I LOVE PEOPLE%").and_return(nil)
-      expect(lambda{
-        response = Organisation.add_email(fields = CSV.parse('i love people,,,,,,,test@example.org')[0],true)
-        expect(response).to eq "i love people was not found\n"
-      }).not_to raise_error
-    end
-
-    it "should add email to org" do
-      expect(Organisation).to receive(:where).with("UPPER(name) LIKE ? ", "%FRIENDLY%").and_return([@org1])
-      expect(@org1).to receive(:email=).with('test@example.org')
-      expect(@org1).to receive(:save)
-      Organisation.add_email(fields = CSV.parse('friendly,,,,,,,test@example.org')[0],true)
-    end
-
-    it "should add email to org even with case mismatch" do
-      expect(Organisation).to receive(:where).with("UPPER(name) LIKE ? ", "%FRIENDLY%").and_return([@org1])
-      expect(@org1).to receive(:email=).with('test@example.org')
-      expect(@org1).to receive(:save)
-      Organisation.add_email(fields = CSV.parse('friendly,,,,,,,test@example.org')[0],true)
-    end
-
-    it 'should not add email to org when it has an existing email' do
-      @org1.email = 'something@example.com'
-      @org1.save!
-      expect(Organisation).to receive(:where).with("UPPER(name) LIKE ? ", "%FRIENDLY%").and_return([@org1])
-      expect(@org1).not_to receive(:email=).with('test@example.org')
-      expect(@org1).not_to receive(:save)
-      Organisation.add_email(fields = CSV.parse('friendly,,,,,,,test@example.org')[0],true)
-    end
-  end
-  
   describe 'destroy uses acts_as_paranoid' do
     it 'can be recovered' do
       @org1.destroy
